move in-mem cache to a new file, added test case for in-mem cache
